django/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
djangorestframework==3.12.4
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
